Enable building of routes table
Previous design was taking a call, building a route for it, then
executing the handler. That defeats the purpose of routing.

This change allows you to build a routing table by calling:

    add_route($method, $route, $handler)

`$method` is an HTTP method (GET|PUT|POST|etc..)
`$route` is the route (`/`, `/foo/:bar`, ..)
`$handler` is the sub which runs when this route is invoked

Other changes:

1. Always return from a method. If nothing particular is returned,
return $self so that methods can be chained
2. `db => $config->{db} //= {}` isn't very clear/might have
side effects. Moved the `//=` parts into their own paragraph
3. Hint to Perl and the programmer that the value returned by
`build_pattern` is a regex, by returning `qr/$pattern/;`

`example.pl` added to show usage