Added token validation
Extended syntax of add_route to allow for an array of regexes.
If this parameter is used, the number of tokens and the number of
regexes must be equal.
If any regex fails to match a token value, mapper() will fail.
Validation errors carp, so each token is checked before a fail response